.docker/local/requirements.txt
python-jenkins~=1.8.2
requests~=2.32.3